+++ configs/exp/trivial_B.yaml
@@ -15,7 +15,7 @@ model:
   dropout: 0.05
 
 optimizer:
-  learning_rate: 1.0e-4
+  learning_rate: 1.0e-3
   optimizer: adamw
   weight_decay: 1.0e-4
 
@@ -26,7 +26,7 @@ loss:
   gamma: 0.99
   value_weight: 1.0e-4
   entropy_weight: 1.0e-1
-  clip_value: 5.5
+  clip_value: 1.0
 
 replay_buffer:
   buffer_size: 50000
@@ -34,7 +34,7 @@ replay_buffer:
 
 iterations:
   rollouts: 10
-  epochs: 300
+  epochs: 2000
   batches: 100
 
 treesearch:
