--- src/treesearch/__init__.py
@@ -1,1 +1,0 @@
-from .td import TDTreeSearch
